02-college-scorecard-analytics/pyproject.toml
[project]
name = "college-scorecard-analytics"
version = "0.1.0"
description = "College Scorecard Student Outcome Analytics — Predictive modeling of post-graduation earnings using institutional characteristics, with SHAP explainability."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<3.14"
authors = [{ name = "Ruthvik Bandari" }]
keywords = ["college-scorecard", "higher-education", "machine-learning", "shap", "xgboost"]

dependencies = [
    "polars[all]>=1.38",
    "pyarrow>=18.0",
    "duckdb>=1.2",
    "scikit-learn>=1.6",
    "xgboost>=2.1",
    "shap>=0.46",
    "plotly>=6.0",
    "kaleido>=0.4",
    "seaborn>=0.13",
    "matplotlib>=3.9",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "ipykernel>=6.29",
    "nbformat>=5.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF", "N", "D"]
ignore = ["D100", "D104", "RUF001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
